refactor: 이동 경로 구하는 공통 메소드 분리
